refactor: split large test files to meet Codacy size limits
- Split CombatStatsSection.test.tsx (480 lines) into 3 focused files:
  - CombatStatsSection.rendering.test.tsx (component rendering and display tests)
  - CombatStatsSection.interactions.test.tsx (user interactions and field updates)
  - CombatStatsSection.validation.test.tsx (validation, constraints, and edge cases)
- Fixed failing test selectors for calculated values to avoid duplicate text matches
- All 41 CombatStatsSection tests passing with proper test organization
- Completes final Codacy method size violation (439 lines > 50 line limit)
- npm run test:ci now passes with 393 tests and 70% coverage
