+++ src/components/forms/character/__tests__/sections/CombatStatsSection.validation.test.tsx
@@ -1,9 +1,8 @@
 import React from 'react';
-import { render, screen, fireEvent } from '@testing-library/react';
-import userEvent from '@testing-library/user-event';
+import { render, screen } from '@testing-library/react';
 import { CombatStatsSection } from '../../sections/CombatStatsSection';
 
-describe('CombatStatsSection', () => {
+describe('CombatStatsSection - Validation and Edge Cases', () => {
   const mockOnChange = jest.fn();
   const defaultProps = {
     value: {
@@ -33,153 +32,6 @@ describe('CombatStatsSection', () => {
 
   beforeEach(() => {
     jest.clearAllMocks();
-  });
-
-  describe('Component Rendering', () => {
-    it('renders section header with proper title', () => {
-      render(<CombatStatsSection {...defaultProps} />);
-
-      expect(screen.getByText('Combat Statistics')).toBeInTheDocument();
-      expect(screen.getByText(/set your character's combat-related statistics/i)).toBeInTheDocument();
-    });
-
-    it('renders all hit points fields', () => {
-      render(<CombatStatsSection {...defaultProps} />);
-
-      expect(screen.getByLabelText(/maximum hit points/i)).toBeInTheDocument();
-      expect(screen.getByLabelText(/current hit points/i)).toBeInTheDocument();
-      expect(screen.getByLabelText(/temporary hit points/i)).toBeInTheDocument();
-    });
-
-    it('renders armor class and speed fields', () => {
-      render(<CombatStatsSection {...defaultProps} />);
-
-      expect(screen.getByLabelText(/armor class/i)).toBeInTheDocument();
-      expect(screen.getByLabelText(/speed \(feet\)/i)).toBeInTheDocument();
-    });
-
-    it('renders proficiency bonus field', () => {
-      render(<CombatStatsSection {...defaultProps} />);
-
-      expect(screen.getByLabelText(/proficiency bonus/i)).toBeInTheDocument();
-    });
-
-    it('displays calculated values section', () => {
-      render(<CombatStatsSection {...defaultProps} />);
-
-      expect(screen.getByText('Calculated Values')).toBeInTheDocument();
-      expect(screen.getByText('Initiative')).toBeInTheDocument();
-      expect(screen.getByText('Hit Die')).toBeInTheDocument();
-      expect(screen.getByText('Total Level')).toBeInTheDocument();
-      expect(screen.getByText('Prof. Bonus')).toBeInTheDocument();
-    });
-  });
-
-  describe('Field Values and Updates', () => {
-    it('displays current hit points values', () => {
-      render(<CombatStatsSection {...defaultProps} />);
-
-      const maxHpField = screen.getByLabelText(/maximum hit points/i);
-      const currentHpField = screen.getByLabelText(/current hit points/i);
-      const tempHpField = screen.getByLabelText(/temporary hit points/i);
-
-      expect(maxHpField).toHaveValue(10);
-      expect(currentHpField).toHaveValue(10);
-      expect(tempHpField).toHaveValue(0);
-    });
-
-    it('displays combat stats values', () => {
-      render(<CombatStatsSection {...defaultProps} />);
-
-      const acField = screen.getByLabelText(/armor class/i);
-      const speedField = screen.getByLabelText(/speed \(feet\)/i);
-      const profBonusField = screen.getByLabelText(/proficiency bonus/i);
-
-      expect(acField).toHaveValue(12);
-      expect(speedField).toHaveValue(30);
-      expect(profBonusField).toHaveValue(2);
-    });
-
-    it('calls onChange when maximum hit points is changed', () => {
-      render(<CombatStatsSection {...defaultProps} />);
-
-      const maxHpField = screen.getByLabelText(/maximum hit points/i);
-      fireEvent.change(maxHpField, { target: { value: '15' } });
-
-      expect(mockOnChange).toHaveBeenCalledWith({
-        ...defaultProps.value,
-        hitPoints: {
-          ...defaultProps.value.hitPoints,
-          maximum: 15,
-        },
-      });
-    });
-
-    it('calls onChange when current hit points is changed', () => {
-      render(<CombatStatsSection {...defaultProps} />);
-
-      const currentHpField = screen.getByLabelText(/current hit points/i);
-      fireEvent.change(currentHpField, { target: { value: '8' } });
-
-      expect(mockOnChange).toHaveBeenCalledWith({
-        ...defaultProps.value,
-        hitPoints: {
-          ...defaultProps.value.hitPoints,
-          current: 8,
-        },
-      });
-    });
-
-    it('calls onChange when temporary hit points is changed', () => {
-      render(<CombatStatsSection {...defaultProps} />);
-
-      const tempHpField = screen.getByLabelText(/temporary hit points/i);
-      fireEvent.change(tempHpField, { target: { value: '5' } });
-
-      expect(mockOnChange).toHaveBeenCalledWith({
-        ...defaultProps.value,
-        hitPoints: {
-          ...defaultProps.value.hitPoints,
-          temporary: 5,
-        },
-      });
-    });
-
-    it('calls onChange when armor class is changed', () => {
-      render(<CombatStatsSection {...defaultProps} />);
-
-      const acField = screen.getByLabelText(/armor class/i);
-      fireEvent.change(acField, { target: { value: '16' } });
-
-      expect(mockOnChange).toHaveBeenCalledWith({
-        ...defaultProps.value,
-        armorClass: 16,
-      });
-    });
-
-    it('calls onChange when speed is changed', () => {
-      render(<CombatStatsSection {...defaultProps} />);
-
-      const speedField = screen.getByLabelText(/speed \(feet\)/i);
-      fireEvent.change(speedField, { target: { value: '25' } });
-
-      expect(mockOnChange).toHaveBeenCalledWith({
-        ...defaultProps.value,
-        speed: 25,
-      });
-    });
-
-    it('calls onChange when proficiency bonus is changed', () => {
-      render(<CombatStatsSection {...defaultProps} />);
-
-      const profBonusField = screen.getByLabelText(/proficiency bonus/i);
-      fireEvent.change(profBonusField, { target: { value: '3' } });
-
-      expect(mockOnChange).toHaveBeenCalledWith({
-        ...defaultProps.value,
-        proficiencyBonus: 3,
-      });
-    });
   });
 
   describe('Calculated Values Display', () => {
@@ -266,53 +118,6 @@ describe('CombatStatsSection', () => {
       render(<CombatStatsSection {...props} />);
 
       expect(screen.getByText('d8')).toBeInTheDocument();
-    });
-  });
-
-  describe('Helper Text and Modifiers', () => {
-    it('displays constitution modifier in hit points helper text', () => {
-      const props = {
-        ...defaultProps,
-        abilityScores: { ...defaultProps.abilityScores, constitution: 16 },
-      };
-      render(<CombatStatsSection {...props} />);
-
-      // Constitution 16 = +3 modifier
-      expect(screen.getByText(/CON modifier: \+3/i)).toBeInTheDocument();
-    });
-
-    it('displays dexterity modifier in armor class helper text', () => {
-      const props = {
-        ...defaultProps,
-        abilityScores: { ...defaultProps.abilityScores, dexterity: 18 },
-      };
-      render(<CombatStatsSection {...props} />);
-
-      // Dexterity 18 = +4 modifier
-      expect(screen.getByText(/DEX modifier: \+4/i)).toBeInTheDocument();
-    });
-
-    it('handles negative constitution modifier', () => {
-      const props = {
-        ...defaultProps,
-        abilityScores: { ...defaultProps.abilityScores, constitution: 8 },
-      };
-      render(<CombatStatsSection {...props} />);
-
-      // Constitution 8 = -1 modifier
-      expect(screen.getByText(/CON modifier: -1/i)).toBeInTheDocument();
-    });
-
-    it('shows default speed helper text', () => {
-      render(<CombatStatsSection {...defaultProps} />);
-
-      expect(screen.getByText(/default: 30 feet/i)).toBeInTheDocument();
-    });
-
-    it('shows temporary hit points as optional', () => {
-      render(<CombatStatsSection {...defaultProps} />);
-
-      expect(screen.getByText('Optional')).toBeInTheDocument();
     });
   });
 
@@ -412,43 +217,6 @@ describe('CombatStatsSection', () => {
       // Should display defaults for missing values
       expect(screen.getByDisplayValue('0')).toBeInTheDocument(); // temporary HP
       expect(screen.getByDisplayValue('30')).toBeInTheDocument(); // speed default
-    });
-
-    it('handles string input conversion to numbers', async () => {
-      const user = userEvent.setup();
-      render(<CombatStatsSection {...defaultProps} />);
-
-      const maxHpField = screen.getByLabelText(/maximum hit points/i);
-      await user.clear(maxHpField);
-      await user.type(maxHpField, 'abc'); // invalid input
-
-      fireEvent.blur(maxHpField);
-
-      expect(mockOnChange).toHaveBeenCalledWith({
-        ...defaultProps.value,
-        hitPoints: {
-          ...defaultProps.value.hitPoints,
-          maximum: 0, // should convert invalid string to 0
-        },
-      });
-    });
-
-    it('correctly updates current HP max based on maximum HP', () => {
-      const props = {
-        ...defaultProps,
-        value: {
-          ...defaultProps.value,
-          hitPoints: {
-            maximum: 25,
-            current: 20,
-            temporary: 0,
-          },
-        },
-      };
-      render(<CombatStatsSection {...props} />);
-
-      const currentHpField = screen.getByLabelText(/current hit points/i);
-      expect(currentHpField).toHaveAttribute('max', '25');
     });
   });
